splitting code throught project
