+++ src/CodeGenerator.cpp
@@ -10,7 +10,8 @@
 #include "ListNode.hpp"
 #include "type.hpp"
 #include "cpptypes.hpp"
-#include "compiler.hpp"
+#include "Compiler.hpp"
+#include "ReturnPacket.hpp"
 #include "symbol_table_stackX.hpp"
 // #include "symtab.hpp"
