+++ src/Makefile.am
@@ -10,9 +10,9 @@ njnrSOURCES += expressions/expressions.cpp
 njnrSOURCES += functions/function_calls.cpp functions/functions.cpp
 njnrSOURCES += statements/statements.cpp
 njnrSOURCES += include/debug.hpp include/type.hpp include/List.hpp include/cpptypes.hpp
-njnrSOURCES += include/ListNode.hpp include/constant.hpp include/common/symbol_table_stackX.hpp
+njnrSOURCES += include/ListNode.hpp include/Constant.hpp include/common/symbol_table_stackX.hpp
 njnrSOURCES += include/common/symbol_table_stack.hpp include/common/symbol_table.hpp
-njnrSOURCES += include/compiler.hpp include/CodeGenerator.hpp include/common/symbol_table_entry.hpp
+njnrSOURCES += include/Compiler.hpp include/CodeGenerator.hpp include/common/symbol_table_entry.hpp
 njnrSOURCES += include/lex.hpp include/data.hpp include/helper.hpp
 njnrSOURCES += include/operator.hpp include/common/symbol_table_templates.h
 njnrSOURCES += include/common/symbol_table_stack_templates.h
