+++ src/expressions/expressions.cpp
@@ -3,7 +3,8 @@
 #include <cstdio>
 #include <iostream>
 #include "debug.hpp"
-#include "compiler.hpp"
+#include "Compiler.hpp"
+#include "ReturnPacket.hpp"
 #include "njnr.tab.hpp"
 #include "symbol_table_stack.hpp"
 #include "symbol_table_stackX.hpp"
