+++ src/include/common/symbol_table_stackX.hpp
@@ -5,9 +5,11 @@
 #include <string>
 
 #include "type.hpp"                // for njnr::type
+#include "ParameterBinding.hpp"
 #include "symbol_table_stack.hpp"  // for class SymbolTable
 #include "symbol_table_entry.hpp"  // for class S_TableEntry
-#include "compiler.hpp"            // for class Compiler
+#include "Compiler.hpp"            // for class Compiler
+
 namespace njnr
 {
 /**
