+++ src/include/type.hpp
@@ -153,114 +153,5 @@ class funcheadertype
    njnr::type    returntype;  // return type(if any) of fn
    njnr::type    ttype;       // type of fn?(TBD)
 };
-/**
- * @brief Type System Type Abstract Class Interface
- * 
- */
-class TSAbstractType
-{
-   public:
-      TSAbstractType() : typeValue{"AbstractType__NA__()"} {};
-
-      explicit TSAbstractType(const std::string typeValue) :
-                                 typeValue{"AbstractType" + typeValue+"()"} {}
-
-      virtual                     ~TSAbstractType()
-      {
-         report(njnr::logType::debug,
-                "running TSAbstractType() destructor");
-         report(njnr::logType::debug, "typeValue" + typeValue);
-      }
-
-      virtual   const bool             getNumeric()    const noexcept = 0;
-      virtual   const bool                getLVal()    const noexcept = 0;
-
-      virtual   const std::string        toString()    const noexcept = 0;
-      virtual   const std::string    getTypeValue()    const noexcept = 0;
-      virtual   const njnr::type          getType()    const noexcept = 0;
-      virtual TSAbstractType& operator=(const TSAbstractType& t)
-      {
-         if (const_cast<TSAbstractType*>(&t) != this)
-         {
-            this->typeValue = t.typeValue;
-         }
-         return *this;
-      }
-
-   protected:
-      std::string typeValue;
-};
-
-/**
- * @brief Generic lower-level implementation of Abstract Class Interface
- *         TSAbstractType
- * 
- */
-class TSType : public TSAbstractType
-{
-   public:
-      TSType() : TSAbstractType("GenericType__NA__()"),
-                                isNumeric{false},
-                                isLval{false} {};
-
-      explicit TSType(const std::string typeValue) :
-                   TSAbstractType("GenericType" + typeValue + "()"),
-                   isNumeric{false},
-                   isLval{false} {};
-
-
-      TSType(const std::string typeValue,
-             const bool isNumeric,
-             const bool isLval = false) :
-             TSAbstractType("GenericType" + typeValue + "()"),
-             isNumeric{isNumeric},
-             isLval{isLval} {};
-
-      virtual ~TSType()
-      {
-         report(njnr::logType::debug,
-                "running TSType() Destructor");
-         report(njnr::logType::debug, this->toString());
-         report(njnr::logType::debug, "isnumeric" + isNumeric);
-         report(njnr::logType::debug, "isLval" + isLval);
-      }
-
-      virtual const bool getNumeric()          const noexcept
-      {
-         return isNumeric;
-      }
-      virtual const bool getLVal()             const noexcept
-      {
-         return isLval;
-      }
-
-      virtual const std::string toString()     const noexcept
-      {
-         return typeValue;
-      }
-      virtual const std::string getTypeValue() const noexcept
-      {
-         return typeValue;
-      }
-      virtual const njnr::type getType()       const noexcept
-      {
-         return njnr::type::VOID;
-      }
-
-      virtual TSType& operator=(const TSType& t)
-      {
-         if (const_cast<TSType*>(&t) != this)
-         {
-            this->typeValue = t.typeValue;
-            this->isNumeric = t.isNumeric;
-            this->isLval = t.isLval;
-         }
-         return *this;
-      }
-
-   private:
-      bool isNumeric;
-      bool isLval;
-};
 }  // namespace njnr
 #endif  // SRC_INCLUDE_TYPE_HPP_
